\name{dkplot}
\alias{dkplot}
\title{
Generate a delta-K plot from area under the curve (AUC) values across multiple cluster numbers.
}
\description{
This function uses the \code{\link{lattice}} function \code{\link{xyplot}} to generate an delta-K plot from a valid \code{"dk"} class object (see \code{\link{dk-class}}).  
}
\usage{
dkplot(x)
}

\arguments{
	\item{x}{
		a valid "dk" class object (see \code{\link{dk-class}}), normally generated by the \code{\link{deltak}} function.
	}
}

\value{No return value, called for side effects}

\references{
Merged consensus clustering to assess and improve class discovery with microarray data. Simpson TI, Armstrong JD and Jarman AP. BMC Bioinformatics 2010, 11:590.
}

\author{
Dr. T. Ian Simpson
\email{ian.simpson@ed.ac.uk}
}

\seealso{
\code{\link{consmatrix-class}}
}
\examples{
#load up a test cluscomp result
data('testcmr');

#look at the result structure
summary(testcmr);

#calculate all of the AUC values from the \code{cluscomp} result for algorithm 'kmeans'
kmeanscmr <- testcmr[grep('kmeans',names(testcmr))];
acs <- aucs(kmeanscmr);

#calculate all of the delta-K values
dks <- deltak(acs);

#plot the delta-K curve
dkplot(dks);
}