\name{cluscomp}
\alias{agnes_clm}
\alias{auc}
\alias{aucs}
\alias{clrob}
\alias{cluscomp}
\alias{data_check}
\alias{deltak}
\alias{agnes_clmem}
\alias{diana_clmem}
\alias{hclust_clmem}
\alias{kmeans_clmem}
\alias{memrob}
\alias{pam_clmem}
\alias{consmatrix-class}
\alias{memroblist-class}
\alias{memrobmatrix-class}
\alias{mergematrix-class}
\alias{validConsMatrixObject}
\alias{validMemRobListObject}
\alias{validMemRobMatrixObject}
\alias{validMergeMatrixObject}
\title{Main functions}
\description{currently undocumented functions}
\keyword{internal}
